+++ docs/release notes/RELEASES.md
@@ -37,7 +37,9 @@ is to be ignored; v3.0.1 is considered the next production release after v2.03
   - #673 fixed: Site branding could sometimes be bypassed (favicon / error boundaries)
   - #663 Users can now have multiple signin methods (google + email/password, multiple emails...),
     and signin email can be different than displayed contact email.
-  - #671 You can now hide from your calendar feed events that you responded "not going".
+  - #671 You can now hide from your calendar feed events that you responded "not going", and hide events you're not invited to.
+  - #686 UI improvements / tweaks
+  - #631 #684 various improvements to metronome (timing stability, edge case bugs, transition continuity)
 
 ## security hardening
 
@@ -95,7 +97,6 @@ Actions taken:
 https://github.com/thenfour/CafeMarcheDB/milestone/28
 
 - #515 full screen mode for mobile should be treated better
-- user admin page needs "edit profile" fields to be restricted.
 
 ## Migration notes